Add film entity and repository and fix tests

+++ src/main/java/com/example/films/core/service/FilmServiceImpl.java
@@ -29,7 +29,7 @@ public class FilmServiceImpl implements FilmService {
         // copy element so changes of element don't affect list
         FilmDTO copy = FilmDTO.of(filmDTO);
 
-        int index = findIndex(copy.getId());
+        long index = findIndex(copy.getId());
         // not found
         if (index == -1) {
             // add
@@ -37,7 +37,7 @@ public class FilmServiceImpl implements FilmService {
             films.add(copy);
         } else {
             // update
-            films.set(index, copy);
+            films.set(Long.valueOf(index).intValue(), copy);
         }
     }
 
@@ -60,14 +60,14 @@ public class FilmServiceImpl implements FilmService {
                 });
     }
 
-    private int findIndex(int id) {
-        List<Integer> ids = films.stream()
+    private int findIndex(long id) {
+        List<Long> ids = films.stream()
                 .map(FilmDTO::getId)
                 .collect(Collectors.toList());
         return ids.indexOf(id);
     }
 
-    private int getNextId(int id) {
+    private long getNextId(long id) {
         if (id > 0) return id;
 
         return films.size() + 1;
